clang-tidy: readability-parameter-inline-comment
New check to spot a function with args with inline comments, moving the
comments up to above the function, with "@param <paramname>:" prefix.

This also similarly moves an inline comment on a variable, member or
enum constant to the line above. But I have disabled that by commenting
it out.

V2: Fixed to not introduce a spurious // in the middle of a comment.

Change-Id: I2987298fbdbbe2cb1289429e3f494026a01ef72c