AMDGPU: Correct const_index_stride for wave 32
Since there is a single scratch resource descriptor for all shaders, if there is
a wave32 and a wave64 shader (for instance the default for gfx10 VsFs pairs)
then the const_index_stride will be incorrect for wave32 shaders.
This change is a temporary fix to alter the const_index_stride
appropriately. A longer term solution will be determined shortly.

Change-Id: I0a3d4a56beacd8fa25fbd1a0a6c27d880e499f2c